Convert useScrub.types to namespace
